add new algo

+++ datastructures/LinkedList/Platform_Types.h
@@ -52,7 +52,7 @@ typedef unsigned short uint16;
 typedef unsigned long uint32;
 typedef signed char sint8;
 typedef signed short sint16;
-typedef signed int sint32;
+typedef signed long sint32;
 typedef float float32;
 #endif
